+++ CathedralOSApp/Services/GenerationRequestDTO.swift
@@ -257,7 +257,7 @@ struct GenerationResponse: Codable {
     /// The number of credits required for this generation (present on insufficient_credits error).
     let requiredCredits: Int?
     /// The number of credits available to the user (present on insufficient_credits error).
-    let availableCredits: Int?
+    let availableCredits: Double?
     /// The number of credits charged for this generation (present on success).
     let creditCostCharged: Double?
     /// The number of credits remaining after this generation (present on success).
@@ -322,7 +322,7 @@ struct GenerationCostEstimate: Codable {
     /// Phase 3 pricing returns fractional credits (6-decimal precision), so this is Double.
     let estimatedCredits: Double
     /// Current available credits for the authenticated user.
-    let availableCredits: Int
+    let availableCredits: Double
     /// `true` if the user has enough credits to proceed with generation.
     let allowed: Bool
     /// Minimum charge enforced by the selected model regardless of actual token usage.
